refactor: Abstract MaterialCache to class.
