+++ Unity Project/Dungeoneering/Assets/_Libs/DungeonCrawler.Unity.Data/src/MaterialCache.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6fc3b1508e5ad5642a3e09e79267042a
